Reminder fixes p2
- shapeCreate and shapeUpdate have been improved by calling preShape before determining what's considered a connect vs. create. This allows objects like reminders to force a connection under certain conditions
- Can now create and update reminders without issue
- Minor bug fixes to functions used by cudHelpers
- Combined some shape/tools files
