+++ packages/shared/src/validation/models/reminderItem.ts
@@ -1,19 +1,19 @@
 import * as yup from "yup";
-import { description, id, minNumErr, name, opt, req, YupModel, yupObj } from "../utils";
+import { blankToUndefined, description, id, maxStrErr, minNumErr, minStrErr, opt, req, YupModel, yupObj } from "../utils";
 
 const index = yup.number().integer().min(0, minNumErr);
 
 export const reminderItemValidation: YupModel = {
     create: ({ o }) => yupObj({
         id: req(id),
-        name: req(name),
+        name: req(yup.string().transform(blankToUndefined).min(1, minStrErr).max(50, maxStrErr)),
         description: opt(description),
         dueDate: opt(yup.date()),
         index: opt(index),
     }, [], [], o),
     update: ({ o }) => yupObj({
         id: req(id),
-        name: opt(name),
+        name: opt(yup.string().transform(blankToUndefined).min(1, minStrErr).max(50, maxStrErr)),
         description: opt(description),
         dueDate: opt(yup.date()),
         index: opt(index),
