+++ packages/ui/src/utils/shape/models/comment.ts
@@ -1,7 +1,6 @@
 import { Comment, CommentCreateInput, CommentFor, CommentTranslation, CommentTranslationCreateInput, CommentTranslationUpdateInput, CommentUpdateInput } from "@local/shared";
 import { ShapeModel } from "types";
-import { createPrims, createRel, shapeUpdate, updatePrims, updateRel } from "./tools";
-import { updateTranslationPrims } from "./tools/updateTranslationPrims";
+import { createPrims, createRel, shapeUpdate, updatePrims, updateRel, updateTranslationPrims } from "./tools";
 
 export type CommentTranslationShape = Pick<CommentTranslation, "id" | "language" | "text"> & {
     __typename?: "CommentTranslation";
