+++ packages/ui/src/utils/shape/models/meeting.ts
@@ -3,8 +3,7 @@ import { ShapeModel } from "types";
 import { MeetingInviteShape, shapeMeetingInvite } from "./meetingInvite";
 import { OrganizationShape } from "./organization";
 import { ScheduleShape, shapeSchedule } from "./schedule";
-import { createPrims, createRel, shapeUpdate, updatePrims, updateRel } from "./tools";
-import { updateTranslationPrims } from "./tools/updateTranslationPrims";
+import { createPrims, createRel, shapeUpdate, updatePrims, updateRel, updateTranslationPrims } from "./tools";
 
 export type MeetingTranslationShape = Pick<MeetingTranslation, "id" | "language" | "description" | "link" | "name"> & {
     __typename?: "MeetingTranslation";
