+++ packages/ui/src/utils/shape/models/reminder.ts
@@ -13,11 +13,20 @@ export type ReminderShape = Pick<Reminder, "id" | "name" | "description" | "dueD
 export const shapeReminder: ShapeModel<ReminderShape, ReminderCreateInput, ReminderUpdateInput> = {
     create: (d) => ({
         ...createPrims(d, "id", "name", "description", ["dueDate", shapeDate], "index"),
-        ...createRel(d, "reminderList", ["Connect", "Create"], "one", shapeReminderList),
         ...createRel(d, "reminderItems", ["Create"], "many", shapeReminderItem, (r) => ({ ...r, reminder: { id: d.id } })),
+        // Treat as connect when the reminderList has created_at
+        ...createRel(d, "reminderList", ["Connect", "Create"], "one", shapeReminderList, (l) => {
+            if (l.created_at) return { id: l.id };
+            return l;
+        }),
     }),
     update: (o, u, a) => shapeUpdate(u, {
         ...updatePrims(o, u, "id", "name", "description", ["dueDate", shapeDate], "index", "isComplete"),
         ...updateRel(o, u, "reminderItems", ["Create", "Update", "Delete"], "many", shapeReminderItem, (r, i) => ({ ...r, reminder: { id: i.id } })),
+        // Treat as connect when the reminderList has created_at
+        ...updateRel(o, u, "reminderList", ["Connect", "Create", "Disconnect"], "one", shapeReminderList, (l) => {
+            if (l.created_at) return { id: l.id };
+            return l;
+        }),
     }, a),
 };
