+++ packages/ui/src/utils/shape/models/reminderItem.ts
@@ -2,17 +2,17 @@ import { ReminderItem, ReminderItemCreateInput, ReminderItemUpdateInput } from "
 import { ShapeModel } from "types";
 import { createPrims, createRel, shapeDate, shapeUpdate, updatePrims } from "./tools";
 
-export type ReminderItemShape = Pick<ReminderItem, "id" | "name" | "description" | "dueDate" | "index"> & {
+export type ReminderItemShape = Pick<ReminderItem, "id" | "name" | "description" | "dueDate" | "index" | "isComplete"> & {
     __typename?: "ReminderItem";
     reminder: { id: string },
 }
 
 export const shapeReminderItem: ShapeModel<ReminderItemShape, ReminderItemCreateInput, ReminderItemUpdateInput> = {
     create: (d) => ({
-        ...createPrims(d, "id", "name", "description", ["dueDate", shapeDate], "index"),
+        ...createPrims(d, "id", "name", "description", ["dueDate", shapeDate], "index", "isComplete"),
         ...createRel(d, "reminder", ["Connect"], "one"),
     }),
     update: (o, u, a) => shapeUpdate(u, {
-        ...updatePrims(o, u, "id", "name", "description", ["dueDate", shapeDate], "index"),
+        ...updatePrims(o, u, "id", "name", "description", ["dueDate", shapeDate], "index", "isComplete"),
     }, a),
 };
