+++ packages/ui/src/utils/shape/models/routineVersionOutput.ts
@@ -2,8 +2,7 @@ import { RoutineVersionOutput, RoutineVersionOutputCreateInput, RoutineVersionOu
 import { ShapeModel } from "types";
 import { hasObjectChanged } from "../general";
 import { shapeStandardVersion, StandardVersionShape } from "./standardVersion";
-import { createPrims, createRel, shapeUpdate, updatePrims, updateRel } from "./tools";
-import { updateTranslationPrims } from "./tools/updateTranslationPrims";
+import { createPrims, createRel, shapeUpdate, updatePrims, updateRel, updateTranslationPrims } from "./tools";
 
 export type RoutineVersionOutputTranslationShape = Pick<RoutineVersionOutputTranslation, "id" | "language" | "description" | "helpText"> & {
     __typename?: "RoutineVersionOutputTranslation";
