+++ packages/ui/src/utils/shape/models/tools/index.ts
@@ -1,11 +1,4 @@
-export * from "./createOwner";
-export * from "./createPrims";
-export * from "./createRel";
-export * from "./createVersion";
+export * from "./creates";
 export * from "./shapeDate";
-export * from "./shapeUpdate";
-export * from "./updateOwner";
-export * from "./updatePrims";
-export * from "./updateRel";
-export * from "./updateVersion";
+export * from "./updates";
